Improved responsibility segregation between the modules of the lambda
- Delegated the responsibility of checking contract to the module responsible for exposing the Lambda endpoint

- Delegated the responsibility of interfacing with google to the maps.ts module

- Fixed the location of the types in the maps lambda
