+++ backend/lambdas/elasticWriter/types.ts
@@ -3,11 +3,3 @@ export interface LooseObject {
 }
 
 export type ActionType = 'create' | 'update' | 'delete';
-
-export type SearchAddressArgument = {
-  input?: string;
-};
-
-export type GetCoordinatesArgument = {
-  placeId?: string;
-};
